recipe_master_api/requirements.txt
appdirs==1.4.3
asgiref==3.6.0
diff-match-patch==20230430
Django==3.2.9
django-cors-headers==4.0.0
django-filter==23.2
django-import-export==1.0.0
django-js-asset==1.2.3
django-mptt==0.9.0
django-rest-framework==0.1.0
django-storages==1.5.2
django-tinymce==3.6.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==19.7.1
olefile==0.44
packaging==16.8
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-decouple==3.0
python3-memcached==1.51
pytz==2016.10
six==1.10.0
sqlparse==0.4.4
tablib==3.4.0
Unidecode==0.4.20
virtualenv==15.1.0
whitenoise==3.3.0